requirements.txt
cycler==0.11.0
fonttools==4.28.3
greenlet==1.1.2
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
msgpack==1.0.3
numpy==1.21.4
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
